fix(ios): 根治 conv1d 崩溃——真实权重核对后修正布局与 dtype。mlx-indextts2 权重均为 MLX 原生布局 (C_out,K,C_in)，移除多余转置；MLX conv1d 为 channels-last，输入/输出做 [B,C,T]↔[B,T,C] 翻转；Ops.conv1d/linear 做 dtype 对齐（权重全 F16）；zeroPad/reflectPad/nearest 转回输入 dtype。另修 W2VBert depthwise 补 groups=D、Campplus 直连 conv 统一修复

+++ ios/Sources/MLXIndexTTS2Core/BigVGAN.swift
@@ -102,7 +102,8 @@ public final class BigVGAN {
         for i in 0..<(b * c * t) {
             for r in 0..<times { out[i * times + r] = f[i] }
         }
-        return MLXArray(out, [b, c, t * times])
+        // 保持输入 dtype（CPU 兜底强转 F32 取数，构造后转回）
+        return MLXArray(out, [b, c, t * times]).asType(x.dtype)
     }
 
     /// mel [1,80,T] → wav [1,1, T*256]
